% Generated by roxygen2 (4.0.1): do not edit by hand
\name{autoplot.table}
\alias{autoplot.table}
\title{Plot a contingency table (i.e. confusion matrix) as a heatmap}
\usage{
\method{autoplot}{table}(object, norm = "none", trans = NULL)
}
\arguments{
\item{object}{contingency table, returned by function table()}

\item{norm}{normalisation method: "none", by row ("rows"), or by column ("columns"); can be abbreviated}

\item{trans}{function used to transform the counts in the contingency table}
}
\description{
Plot a contingency table (i.e. confusion matrix) as a heatmap
}
\examples{
n <- 50
pred <- sample(letters[1:5], size=n, replace=TRUE)
valid <- sample(letters[2:8], size=n, replace=TRUE)
x <- confusion_matrix(pred, valid)
# autoplot(x)
}
\seealso{
Other confusion.statistics.functions: \code{\link{confusion_matrix}};
  \code{\link{confusion_stats}}
}

