% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bin_particles_profiles}
\alias{bin_particles_profiles}
\title{Bin UVP profiles over depth}
\usage{
bin_particles_profiles(x, bin = 1)
}
\arguments{
\item{x}{a data.frame resulting from \code{\link{read_particles_profile}}, with one or several UVP profiles}

\item{bin}{the size of the depth bin in m}
}
\value{
A data.frame similar to the input data.frame but binned in depth
}
\description{
Bin UVP profiles over depth
}
\details{
Average the concentration, esd, and mean grey level over depth bins
}

