% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/confusion.R
\name{blank_diagonal}
\alias{blank_diagonal}
\title{Blank-out the diagonal of a confusion matrix}
\usage{
blank_diagonal(x)
}
\arguments{
\item{x}{confusion matrix (as output by \code{\link{confusion_matrix}})}
}
\value{
The confusion matrix with NAs on the diagonal.
}
\description{
Blank-out the diagonal of a confusion matrix
}
\details{
Blanking-out the diagonal allows to better focus on the prediction errors, particularly for plots.
}
\examples{
n <- 50
pred <- sample(letters[1:5], size=n, replace=TRUE)
valid <- sample(letters[2:8], size=n, replace=TRUE)
m <- confusion_matrix(pred, valid)
m
blank_diagonal(m)
\dontrun{
library("ggplot2")
autoplot(blank_diagonal(m))
}
}
\seealso{
Other confusion statistics functions: \code{\link{autoplot.table}};
  \code{\link{confusion_matrix}};
  \code{\link{confusion_stats}}
}

