% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_project.R
\name{check}
\alias{check}
\title{Check consistency of data in a project}
\usage{
check(project, nmax = 50)
}
\arguments{
\item{project}{path to the project directory}

\item{nmax}{maximum number of problematic items to print out (50 by default)}
}
\value{
Returns a list with all problematic items (sample or object names), invisibly.
}
\description{
Check consistency of data in a project
}
\details{
This function
\enumerate{
  \item Reads the metadata table of the project
  \item Lists and reads all \code{.pid} files in the \code{work} directory
  \item Lists and reads all \code{dat1.tx} files in the \code{PID_process/Pid_results/Dat1_validated} directory
  \item Lists all images in the \code{PID_process/Sorted_vignettes} directory
}
Then it
\enumerate{
  \item Compares the list of samples in the metadata, \code{.pid}, and \code{.txt} files
  \item Compares the list of objects in the \code{.pid}, \code{.txt} files, and present as images (i.e. vignettes)
  \item For images present both in the \code{.txt} files and as vignettes, checks that the location of a vignette (its classification) matches the last column in the \code{txt} file
}
}

