% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check.R
\name{check_project}
\alias{check_project}
\alias{print.checklist}
\title{Check consistency of data in a project}
\usage{
check_project(project)

\method{print}{checklist}(x, nmax = 50)
}
\arguments{
\item{project}{path to the project directory}

\item{nmax}{maximum number of problematic items to print out (50 by default)}
}
\value{
Returns a list (of class \code{check}) containing all problematic items (sample or object names) that can be printed with various levels of detail (depending on \code{nmax})
}
\description{
Check consistency of data in a project
}
\details{
This function
\enumerate{
  \item Reads the metadata table of the project
  \item Lists and reads all \code{dat1.pid} files in the \code{work} directory
  \item Lists and reads all \code{dat1.tx} files in \code{PID_process/Pid_results/Dat1_validated} (or \code{Dat1_extracted} if the previous one is empty)
  \item Lists all images in the \code{PID_process/Sorted_vignettes} directory
}
Then it
\enumerate{
  \item Compares the list of samples in the metadata, \code{dat1.pid}, and \code{dat1.txt} files
  \item Compares the list of objects in the \code{dat1.pid}, \code{dat1.txt} files, and present as images (i.e. vignettes)
  \item For images present both in the \code{dat1.txt} files and as vignettes, checks that the location of a vignette (its classification) matches the last column in the \code{dat1.txt} file
}
}

