% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/confusion.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Compute a confusion matrix}
\usage{
confusion_matrix(pred, valid)
}
\arguments{
\item{pred}{vector of predicted categories}

\item{valid}{vector of true categories, after validation}
}
\value{
An object of class \code{\link[base]{table}} with the predicted categories as rows and the true, validated categories as columns. The list of categories is made the same for each which always result in a square matrix
}
\description{
Compute a confusion matrix
}
\examples{
n <- 50
pred <- sample(letters[1:5], size=n, replace=TRUE)
valid <- sample(letters[2:8], size=n, replace=TRUE)
confusion_matrix(pred, valid)
}
\seealso{
Other confusion statistics functions: \code{\link{autoplot.table}};
  \code{\link{confusion_stats}}
}

