% Generated by roxygen2 (4.0.2): do not edit by hand
\name{project_type}
\alias{project_type}
\title{Determine the type of project (zooscan or uvp)}
\usage{
project_type(project)
}
\arguments{
\item{project}{path to the project directory}
}
\value{
A character string with the project type, all in lowercase
}
\description{
Determine the type of project (zooscan or uvp)
}
\details{
Extracts the first part of the name of the project which should be either Zooscan or uvp5.
}

