% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_ids}
\alias{read_ids}
\title{Read the identifications of all objects in a project}
\usage{
read_ids(project, last = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{project}{path to the project directory}

\item{last}{when TRUE, only output the prediction and the last identification; when false, output all identifications (but still label the last one "Valid")}

\item{verbose}{when TRUE, show progress if needed}

\item{...}{passed to \code{\link{read_pid}}}
}
\description{
Read the identifications of all objects in a project
}

