% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_learning_set.R
\name{read_learning_set}
\alias{read_learning_set}
\title{Read the learning set of a project}
\usage{
read_learning_set(project)
}
\arguments{
\item{project}{path to the project directory}
}
\description{
Read the learning set of a project
}
\seealso{
Other project data handling functions: \code{\link{read_ids}},
  \code{\link{read_meta}},
  \code{\link{read_particles_profiles}},
  \code{\link{read_variables}}
}

