% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_learning_set_in_project}
\alias{read_learning_set_in_project}
\title{Read the learning set of a project}
\usage{
read_learning_set_in_project(project)
}
\arguments{
\item{project}{path to the project directory}
}
\description{
Walk the hierachy of a zooprocess project and read the learning set file
}

