% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_particles_profile}
\alias{read_particles_profile}
\title{Read the profile of particles in a UVP project}
\usage{
read_particles_profile(project, verbose = TRUE, n = 21, ...)
}
\arguments{
\item{project}{path to the project directory}

\item{verbose}{when TRUE, show read progress (when there are many files to read)}

\item{n}{number of images to aggregate data over. The higher the smoother the profiles.}

\item{...}{passed to \code{\link{read.table}}}
}
\description{
This reads the characteristics of particles in each UVP image (in results/***_datfile.txt), reads the appropriate metadata (in meta), and aggregates the data over a given number of images to smooth out noise.
}
\seealso{
Other project data handling functions: \code{\link{read_ids}};
  \code{\link{read_learning_set_in_project}};
  \code{\link{read_meta}}; \code{\link{read_sample_meta}};
  \code{\link{read_variables}}
}

