% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_particles_profiles}
\alias{read_particles_profiles}
\title{Read the profile of particles in a UVP project}
\usage{
read_particles_profiles(project, ...)
}
\arguments{
\item{project}{path to the project directory}

\item{...}{passed to \code{\link{read.table}}}
}
\description{
This reads the characteristics of particles in each UVP image (in results/***_datfile.txt), reads the appropriate metadata (in meta), and aggregates the data over a given number of images to smooth out noise.
}
\seealso{
\code{\link{smooth_particles_profiles}} to smooth the noisy raw signal (through a moving average with function \code{\link{mav}})

Other project data handling functions: \code{\link{read_ids}};
  \code{\link{read_learning_set_in_project}};
  \code{\link{read_meta}}; \code{\link{read_variables}}
}

