% Generated by roxygen2 (4.0.0): do not edit by hand
\name{read_pid}
\alias{read_pid}
\title{Read .pid (or _dat1.txt) files}
\usage{
read_pid(file, data = TRUE, metadata = FALSE)
}
\arguments{
\item{file}{path to a \code{.pid} or \code{dat1.txt} file}

\item{data}{boolean, wether to extract the data table}

\item{metadata}{boolean, wether to extract the metadata from the header}
}
\value{
A data.frame, possibly with a meta attribute containing the metadata as a named list; or the list of metadata directly, if only this is requested
}
\description{
Read .pid (or _dat1.txt) files
}

