% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_pids_in_project}
\alias{read_pids_in_project}
\title{Read all .pid files in a project}
\usage{
read_pids_in_project(project, ...)
}
\arguments{
\item{project}{path to the project directory}
}
\description{
Walk the hierarchy of a zooprocess project and read the pid + dat1.txt files for all records in the project merging the information in both
\itemize{
 \item pid = image characteristics
 \item dat1 = identification
}
}

