% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_variables}
\alias{read_variables}
\title{Read the descriptive variables of all objects in a project}
\usage{
read_variables(project, verbose = TRUE, ...)
}
\arguments{
\item{project}{path to the project directory}

\item{verbose}{when TRUE, show progress if needed}

\item{...}{passed to \code{\link{read_pid}}}
}
\description{
Read the descriptive variables of all objects in a project
}

