% Generated by roxygen2 (4.0.1): do not edit by hand
\name{regroup}
\alias{regroup}
\title{Regroup some categories together}
\usage{
regroup(x, matches)
}
\arguments{
\item{x}{vector of categories (usually the "Valid" column of the data.frame produced by \code{\link{read_ids}})}

\item{matches}{a two columns data.frame or matrix with the old names in the first column and the new name in the second column; if some new names are equal, this effectively regroups the categories}
}
\description{
Regroup some categories together
}
\examples{
x <- c("append_fritill", "append_housing", "crust_copepods_calanus", "crust_copepods_euchaeta", "crust_larvae", "append_fritill")
grouping <- data.frame(old=c("append_fritill", "append_housing"), new="appendicularians")
print(grouping)
regroup(x, matches=grouping)
}

