% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regroup.R
\name{regroup_taxa}
\alias{regroup_taxa}
\title{Regroup some categories along a taxonomy}
\usage{
regroup_taxa(x, level = 1, pattern = NULL, separator = "_")
}
\arguments{
\item{x}{vector of categories (usually the "Valid" column of the data.frame produced by \code{\link{read_ids}})}

\item{level}{level at which to regroup (1 is what precedes the first separator, 2 is what precedes the second separator, etc.)}

\item{pattern}{when provided, only act on categories whose name matches the pattern (matching is done by \code{stringr::str_detect})}

\item{separator}{character used to separate taxonomic levels in category names; usually "_"}
}
\description{
Categories are usually labelled something like "crust_copepods_calanus", "crust_copepods_euchaeta", "crust_larvae", etc.; with different taxonomic levels separated by "_". This function regroups categories in higher level taxonomic units (such as "crust" here).
}
\examples{
x <- c("append_fritill", "append_housing", "crust_copepods_calanus", "crust_copepods_euchaeta", "crust_larvae")
regroup_taxa(x)
regroup_taxa(x, level=2)
regroup_taxa(x, pattern="crust")
}

