% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regroup.R
\name{relabel}
\alias{relabel}
\title{Relabel categories, possibly regrouping them}
\usage{
relabel(x, labels)
}
\arguments{
\item{x}{vector of categories (usually the "Valid" column of the data.frame produced by \code{\link{read_ids}})}

\item{labels}{a two columns data.frame or matrix with the old names in the first column and the new name in the second column; if some new names are equal, this effectively regroups the categories}
}
\description{
Relabel categories, possibly regrouping them
}
\examples{
x <- c("append_fritill", "append_housing", "crust_copepods_calanus", "crust_copepods_euchaeta", "crust_larvae", "append_fritill")

# change labels
labels <- data.frame(old=c("append_fritill", "append_housing"), new=c("larvacean_fritill", "larvacean_housing"))
relabel(x, labels=labels)

# combine some groups (see regroup_taxa for a more generic way of doing this)
grouping <- data.frame(old=c("append_fritill", "append_housing"), new="appendicularians")
print(grouping)
relabel(x, labels=grouping)

}
\seealso{
\code{\link{regroup_taxa}}
}

