% Generated by roxygen2 (4.0.2): do not edit by hand
\name{smooth_particles_profiles}
\alias{smooth_particles_profiles}
\title{Smooth small scale variability in UVP profiles}
\usage{
smooth_particles_profiles(x, n = 21)
}
\arguments{
\item{x}{a data.frame resulting from \code{\link{read_particles_profiles}}, with one or several UVP profiles}

\item{n}{number of images to smooth data over. The higher, the smoother the profiles.}
}
\value{
A data.frame identical to the input data.frame but with smoothed data
}
\description{
Smooth small scale variability in UVP profiles
}
\details{
Compute a moving average of \code{concentration}, \code{esd}, and \code{mean_grey} with function \code{link{mav}}
}

