% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.R
\name{autoplot.table}
\alias{autoplot.table}
\title{Plot a contingency table (i.e. confusion matrix) as a heat map}
\usage{
\method{autoplot}{table}(object, norm = "none", trans = NULL)
}
\arguments{
\item{object}{contingency table, returned by function \code{table()} (or \code{\link{confusion_matrix}})}

\item{norm}{normalisation method: "none", by row ("rows", to represent precision), or by column ("columns", to represent recall); can be abbreviated}

\item{trans}{function used to transform the counts in the contingency table (such as \code{sqrt}, \code{log}, \code{log1p})}
}
\description{
Plot a contingency table (i.e. confusion matrix) as a heat map
}
\examples{
n <- 50
pred <- sample(letters[1:5], size=n, replace=TRUE)
valid <- sample(letters[2:8], size=n, replace=TRUE)
x <- confusion_matrix(pred, valid)
# autoplot(x)
}
\seealso{
Other confusion statistics functions: \code{\link{blank_diagonal}},
  \code{\link{confusion_matrix}},
  \code{\link{confusion_stats}}
}

