% Generated by roxygen2 (4.0.1): do not edit by hand
\name{confusion_stats}
\alias{confusion_stats}
\title{Compute confusion statistics (recall, precision, etc.)}
\usage{
confusion_stats(x, sort.by = NULL)
}
\arguments{
\item{x}{confusion matrix (table class with prediction as line and observations as columns), as returned by \code{confusion_matrix()}}

\item{sort.by}{column to sort the result by (usually "recall", "precision", or "F1")}
}
\description{
Compute confusion statistics (recall, precision, etc.)
}
\seealso{
Other confusion.statistics.functions: \code{\link{autoplot.table}};
  \code{\link{confusion_matrix}}
}

