% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_particles_profiles.R
\name{mav}
\alias{mav}
\title{Compute a moving average}
\usage{
mav(x, n)
}
\arguments{
\item{x}{a numeric vector}

\item{n}{size of the window over which to compute the moving average}
}
\value{
A numeric vector of the same length as x, with the moving average values
}
\description{
Compute a moving average
}
\details{
The function computes a weighted, centred moving average: each data point is computed as the average of the points around it, with decreasing weights as one moves away from the centre. The original data is padded on the left and on the right to be able to compute moving average values at the extremities.
}

