% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_type.R
\name{project_type}
\alias{project_type}
\title{Determine the type of project (zooscan or uvp)}
\usage{
project_type(project)
}
\arguments{
\item{project}{path to the project directory}
}
\value{
A character string with the project type, all in lowercase
}
\description{
Determine the type of project (zooscan or uvp)
}
\details{
Extracts the first part of the name of the project which should be either Zooscan or uvp5.
}

