% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_learning_set_in_project}
\alias{read_learning_set_in_project}
\title{Read the learning set of a project}
\usage{
read_learning_set_in_project(project)
}
\arguments{
\item{project}{path to the project directory}
}
\description{
Read the learning set of a project
}
\seealso{
Other project data handling functions: \code{\link{read_ids}};
  \code{\link{read_meta}}; \code{\link{read_variables}}
}

