% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_meta}
\alias{read_meta}
\title{Read the metadata file for a project}
\usage{
read_meta(project)
}
\arguments{
\item{project}{path to the project directory}
}
\value{
A data.frame with data appropriate for the type of project.
}
\description{
Read the metadata file for a project
}
\details{
First guess the type of project from its name. If the project is a UVP project, read the file in \code{meta} and return the result. If the project is a Zooscan project, read the two files in \code{Zooscan_meta} (one for the samples and one for the scans), join the information in a single data.frame, and return it.
}
\seealso{
Other project data handling functions: \code{\link{read_ids}};
  \code{\link{read_learning_set_in_project}};
  \code{\link{read_particles_profile}};
  \code{\link{read_sample_meta}};
  \code{\link{read_variables}}
}

