% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_pid}
\alias{read_pid}
\title{Read dat1.pid or dat1.txt files}
\usage{
read_pid(file, data = TRUE, metadata = FALSE, verbose = FALSE)
}
\arguments{
\item{file}{path to a \code{dat1.pid} or \code{dat1.txt} file}

\item{data}{boolean, wether to extract the data table}

\item{metadata}{boolean, wether to extract the metadata from the header}
}
\value{
A data.frame, possibly with an attribute named \code{meta} containing the metadata as a named list; or the list of metadata directly, if only metadata is requested
}
\description{
Read dat1.pid or dat1.txt files
}

