% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_variables}
\alias{read_variables}
\title{Read the descriptive variables of all objects in a project}
\usage{
read_variables(project, ...)
}
\arguments{
\item{project}{path to the project directory}

\item{...}{passed to \code{\link{read_pid}}}
}
\description{
Read the descriptive variables of all objects in a project
}
\seealso{
Other project data handling functions: \code{\link{read_ids}};
  \code{\link{read_learning_set}}; \code{\link{read_meta}};
  \code{\link{read_particles_profiles}}
}

